EVA HMA Market.md
The global EVA HMA market size was estimated at USD 482 million in 2023 and is projected to reach USD 747.74 million by 2032, exhibiting a CAGR of 5.00% during the forecast period. In North America, the EVA HMA market size was estimated at USD 136.59 million in 2023, with a CAGR of 4.29% during the forecast period of 2024 through 2032.