Extensive Green-Roof Market.md
The global Extensive Green-Roof market size was estimated at USD 1507 million in 2023 and is projected to reach USD 2546.04 million by 2032, exhibiting a CAGR of 6.00% during the forecast period. North America Extensive Green-Roof market size was estimated at USD 434.11 million in 2023, at a CAGR of 5.14% during the forecast period of 2024 through 2032.