Extruded Nylon Market.md
The global Extruded Nylon market size was estimated at USD 1154 million in 2023 and is projected to reach USD 1949.66 million by 2032, exhibiting a CAGR of 6.00% during the forecast period. North America Extruded Nylon market size was estimated at USD 332.42 million in 2023, at a CAGR of 5.14% during the forecast period of 2024 through 2032.