Fermented Plant Extract Market.md
The global Fermented Plant Extract market size was estimated at USD 983.40 million in 2023 and is projected to reach USD 2049.26 million by 2032, exhibiting a CAGR of 8.50% during the forecast period. In North America, the market size was estimated at USD 294.94 million in 2023, with a CAGR of 7.29% during the forecast period of 2024 through 2032.