Building and Construction Expanded Polystyrene (EPS) Market.md
The global Building and Construction Expanded Polystyrene (EPS) market was estimated at USD 1536.60 million in 2023, with a projected growth to reach USD 2168.21 million by 2032, demonstrating a CAGR of 3.90% during the forecast period. North America's market size for Building and Construction Expanded Polystyrene (EPS) was estimated at USD 427.61 million in 2023, showing a CAGR of 3.34% from 2024 to 2032.