Fermentation Chemical Market.md
The global Fermentation Chemical market size was estimated at USD 147490 million in 2023 and is projected to reach USD 185818.45 million by 2032, with a CAGR of 2.60% during the forecast period. North America's market size for Fermentation Chemical was estimated at USD 40163.72 million in 2023, with a CAGR of 2.23% during the forecast period from 2024 to 2032.