Ethyl Lactate for Semiconductors Market.md
The global Ethyl Lactate for Semiconductors market size was estimated at USD 24.70 million in 2023 and is projected to reach USD 44.28 million by 2032, exhibiting a CAGR of 6.70% during the forecast period. North America Ethyl Lactate for Semiconductors market size was estimated at USD 7.20 million in 2023, at a CAGR of 5.74% during the forecast period of 2024 through 2032.