Ectoine Market.md
Ectoine, chemically known as tetrahydromethymidine carboxylic acid/tetrahydropyridine, is a derivative of an amino acid that originated in a salt lake in Egypt. The molecules of ectoine, which are highly absorbable and structurally special, act as "little shields" around bacteria's DNA and proteins, reducing the exposure to intense UV rays in high-salt conditions. Its unique molecular structure gives it strong water molecular complexing ability.