EDTA Sodium Salt Market.md
The global EDTA Sodium Salt market size was estimated at USD 537 million in 2023 and is projected to reach USD 854.73 million by 2032, exhibiting a CAGR of 5.30% during the forecast period. The North America EDTA Sodium Salt market size was estimated at USD 152.93 million in 2023, at a CAGR of 4.54% during the forecast period of 2024 through 2032.