Earth Friendly Plastic Bags Market.md
The global Earth Friendly Plastic Bags market size was estimated at USD 1684.50 million in 2023 and is projected to reach USD 2590.90 million by 2032, with a CAGR of 4.90% during the forecast period. North America's Earth Friendly Plastic Bags market size in 2023 was estimated at USD 476.58 million, with a CAGR of 4.20% for the forecast period of 2024 through 2032.