Enriched BF3 Market.md
The global Enriched BF3 market size was estimated at USD 101 million in 2023 and is projected to reach USD 169.19 million by 2032, exhibiting a CAGR of 5.90% during the forecast period. North America Enriched BF3 market size was estimated at USD 29.05 million in 2023, at a CAGR of 5.06% during the forecast period of 2024 through 2032.